tz.ptvo_switch_trigger, tz.ptvo_switch_uart, tz.ptvo_switch_analog_input and tz.ptvo_switch_light_brightness
