+++ src/converters/toZigbee.ts
@@ -2250,93 +2250,6 @@ export const humidity: Tz.Converter = {
 // #endregion
 
 // #region Non-generic converters
-export const ptvo_switch_trigger: Tz.Converter = {
-    key: ["trigger", "interval"],
-    convertSet: async (entity, key, value, meta) => {
-        utils.assertNumber(value, key);
-        utils.assertEndpoint(entity);
-        if (key === "trigger") {
-            await entity.command("genOnOff", "onWithTimedOff", {ctrlbits: 0, ontime: Math.round(value / 100), offwaittime: 0});
-        } else if (key === "interval") {
-            const cluster = "genOnOff";
-            if (entity.supportsInputCluster(cluster) || entity.supportsOutputCluster(cluster)) {
-                await entity.configureReporting(cluster, [
-                    {
-                        attribute: "onOff",
-                        minimumReportInterval: value,
-                        maximumReportInterval: value,
-                        reportableChange: 0,
-                    },
-                ]);
-            } else if (utils.hasEndpoints(meta.device, [1])) {
-                const endpoint = meta.device.getEndpoint(1);
-                await endpoint.configureReporting("genBasic", [
-                    {
-                        attribute: "zclVersion",
-                        minimumReportInterval: value,
-                        maximumReportInterval: value,
-                        reportableChange: 0,
-                    },
-                ]);
-            }
-        }
-    },
-};
-export const ptvo_switch_uart: Tz.Converter = {
-    key: ["action"],
-    convertSet: async (entity, key, value, meta) => {
-        if (!value) {
-            return;
-        }
-        const payload = {14: {value, type: 0x42}};
-        for (const endpoint of meta.device.endpoints) {
-            const cluster = "genMultistateValue";
-            if (endpoint.supportsInputCluster(cluster) || endpoint.supportsOutputCluster(cluster)) {
-                await endpoint.write(cluster, payload);
-                return;
-            }
-        }
-        await entity.write("genMultistateValue", payload);
-    },
-};
-export const ptvo_switch_analog_input: Tz.Converter = {
-    key: ["l1", "l2", "l3", "l4", "l5", "l6", "l7", "l8", "l9", "l10", "l11", "l12", "l13", "l14", "l15", "l16"],
-    convertGet: async (entity, key, meta) => {
-        const epId = Number.parseInt(key.substr(1, 2), 10);
-        if (utils.hasEndpoints(meta.device, [epId])) {
-            const endpoint = meta.device.getEndpoint(epId);
-            await endpoint.read("genAnalogInput", ["presentValue", "description"]);
-        }
-    },
-    convertSet: async (entity, key, value, meta) => {
-        const epId = Number.parseInt(key.substr(1, 2), 10);
-        if (utils.hasEndpoints(meta.device, [epId])) {
-            const endpoint = meta.device.getEndpoint(epId);
-            let cluster = "genLevelCtrl";
-            if (endpoint.supportsInputCluster(cluster) || endpoint.supportsOutputCluster(cluster)) {
-                const value2 = Number(value);
-                if (Number.isNaN(value2)) {
-                    return;
-                }
-                const payload = {currentLevel: value2};
-                await endpoint.write(cluster, payload);
-                return;
-            }
-
-            cluster = "genAnalogInput";
-            if (endpoint.supportsInputCluster(cluster) || endpoint.supportsOutputCluster(cluster)) {
-                const value2 = Number(value);
-                if (Number.isNaN(value2)) {
-                    return;
-                }
-                const payload = {presentValue: value2};
-                await endpoint.write(cluster, payload);
-                return;
-            }
-        }
-        return;
-    },
-};
 export const tint_scene: Tz.Converter = {
     key: ["tint_scene"],
     convertSet: async (entity, key, value, meta) => {
@@ -2738,38 +2651,3 @@ export const ignore_rate: Tz.Converter = {
     convertSet: async (entity, key, value, meta) => {},
 };
 // #endregion
-
-export const ptvo_switch_light_brightness: Tz.Converter = {
-    key: ["brightness", "brightness_percent", "transition"],
-    options: [exposes.options.transition()],
-    convertSet: async (entity, key, value, meta) => {
-        if (key === "transition") {
-            return;
-        }
-        const cluster = "genLevelCtrl";
-        utils.assertEndpoint(entity);
-        if (entity.supportsInputCluster(cluster) || entity.supportsOutputCluster(cluster)) {
-            const message = meta.message;
-
-            let brightness: number;
-            if (message.brightness != null) {
-                brightness = Number(message.brightness);
-            } else if (message.brightness_percent != null) brightness = Math.round(Number(message.brightness_percent) * 2.55);
-
-            if (brightness !== undefined && brightness === 0) {
-                message.state = "off";
-                message.brightness = 1;
-            }
-            return await light_onoff_brightness.convertSet(entity, key, value, meta);
-        }
-        throw new Error("LevelControl not supported on this endpoint.");
-    },
-    convertGet: async (entity, key, meta) => {
-        const cluster = "genLevelCtrl";
-        utils.assertEndpoint(entity);
-        if (entity.supportsInputCluster(cluster) || entity.supportsOutputCluster(cluster)) {
-            return await light_onoff_brightness.convertGet(entity, key, meta);
-        }
-        throw new Error("LevelControl not supported on this endpoint.");
-    },
-};
