+++ src/devices/xyzroe.ts
@@ -1,5 +1,4 @@
 import * as fz from "../converters/fromZigbee";
-import * as tz from "../converters/toZigbee";
 import * as ptvo from "../devices/custom_devices_diy";
 import * as exposes from "../lib/exposes";
 import * as m from "../lib/modernExtend";
@@ -376,7 +375,12 @@ export const definitions: DefinitionWithExtend[] = [
             ptvo.fzLocal.ptvo_multistate_action,
             fzLocal.zigusb_button_config,
         ],
-        toZigbee: [tzLocal.zigusb_restart_interval, tzLocal.zigusb_on_off_invert, tz.ptvo_switch_analog_input, tzLocal.zigusb_button_config],
+        toZigbee: [
+            tzLocal.zigusb_restart_interval,
+            tzLocal.zigusb_on_off_invert,
+            ptvo.tzLocal.ptvo_switch_analog_input,
+            tzLocal.zigusb_button_config,
+        ],
         exposes: [
             e.switch().withEndpoint("l1"),
             e
